# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Parking API Documentation",
    "version": "1.0.0",
    "description": " API documentation for your Express application"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API server"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/auth/registers": {
      "get": {
        "summary": "Fetch registration details",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Registration details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "The user ID"
                    },
                    "firstname": {
                      "type": "string",
                      "description": "The user's first name"
                    },
                    "lastname": {
                      "type": "string",
                      "description": "The user's last name"
                    },
                    "email": {
                      "type": "string",
                      "description": "The user's email address"
                    }
                  }
                },
                "example": {
                  "id": "12345",
                  "firstname": "John",
                  "lastname": "Doe",
                  "email": "johndoe@example.com"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing or invalid parameters)"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "The user's email"
                  },
                  "password": {
                    "type": "string",
                    "description": "The user's password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "auth/{id}": {
      "delete": {
        "summary": "Delete a user by ID",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "400": {
            "description": "Bad request (e.g., invalid ID format)"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "description": "The user's name"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The user's email"
          },
          "password": {
            "type": "string",
            "description": "The user's password (plain text for registration, to be hashed)"
          },
          "address": {
            "type": "string",
            "description": "The user's address"
          },
          "access_level": {
            "type": "string",
            "enum": [
              "Admin(Full)",
              "Receptionist(moderate)"
            ],
            "description": "The user's role"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "format": "objectId",
            "description": "User who created the record"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "format": "objectId",
            "description": "User who last updated the record"
          },
          "isBlocked": {
            "type": "boolean",
            "default": false
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "NewUser": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "description": "The user's name"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The user's email"
          },
          "password": {
            "type": "string",
            "description": "The user's password (plain text for registration, to be hashed)"
          },
          "address": {
            "type": "string",
            "description": "The user's address"
          },
          "access_level": {
            "type": "string",
            "enum": [
              "Admin(Full)",
              "Receptionist(moderate)"
            ],
            "description": "The user's role"
          }
        }
      },
...
    }
  }
}